Added meme captioning

+++ src/commands/memify.js
@@ -5,9 +5,11 @@ import hash from 'object-hash';
 import request from 'request-promise';
 import path from 'path';
 import uuidv4 from 'uuid';
+import winston from 'winston';
 
 import CaptionedImage from 'commands/CaptionedImage';
 import Caption from 'models/Caption';
+import Config from 'config';
 import Meme from 'models/Meme';
 import SearchResults from 'models/SearchResults';
 
@@ -22,37 +24,44 @@ const Phase = {
 };
 
 class Memify {
-  promise: Promise<Meme>;
   meme: Meme;
   message: Object;
 
   constructor(message: Object) {
     this.message = message;
-    this.meme = Meme.forge({ user: message.user, team: message.team })
-    this.promise = this.meme.orderBy('updated_at', 'DESC').fetch();
+    this.meme = this.getMeme();
+  }
+
+  async getMeme() {
+    let meme = await Meme.forge({ user: this.message.user, team: this.message.team })
+      .orderBy('updated_at', 'DESC').fetch();
+    if (meme) {
+      if (meme.get('phase') === Phase.DONE && this.message.text.startsWith('search "')) {
+        meme = null;
+      }
+    }
+
+    if (!meme) {
+      meme = await Meme.forge({
+        user: this.message.user,
+        team: this.message.team,
+        phase: Phase.QUERY
+      }).save();
+    }
+    return meme;
   }
 
   async doThing() {
-    let meme = await this.promise;
-    if (meme) {
-      this.meme = meme;
-    } else {
-      // Generate an id.
-      this.meme = await this.meme.save();
-    }
+    this.meme = await this.meme;
     let phase = this.meme.get('phase');
-    if (!phase || this.message.text.startsWith('search "')) {
-      // New query.
-      phase = Phase.QUERY;
-      this.meme = await Meme.forge({ user: this.message.user, team: this.message.team }).save();
-    }
+    winston.info(`phase ${phase}: ${this.message.text}`);
     switch (phase) {
       case Phase.QUERY:
-        return this.searchForImages();
+        return this.newSearch();
       case Phase.SEARCH_RESULTS:
-        const text = this.message.text;
-        if (text.trim() === 'next') {
-          return this.next();
+        const text = this.message.text.trim();
+        if (text === 'next' || text === 'previous') {
+          return this.nav(text);
         }
         // Otherwise, fallthrough.
       case Phase.CHOOSE_IMAGE:
@@ -64,9 +73,12 @@ class Memify {
     }
   }
 
-  searchForImages() {
-    const query = this.message.text;
-    const results = SearchResults.search(query, this.meme.id);
+  async newSearch() {
+    let query = this.message.text;
+    if (query.startsWith('search "')) {
+      query = query.substring('search "'.length, query.lastIndexOf('"'));
+    }
+    const results = await SearchResults.search(query, this.meme.get('id'));
     if (!results.hasNext()) {
       return { text: "No images found" };
     }
@@ -74,10 +86,17 @@ class Memify {
     return results.getAttachment();
   }
 
-  async next() {
-    const results = await SearchResults.forge({ meme_id: this.meme.id }).fetch({ require: true });
-    if (results.hasNext()) {
+  async nav(direction: string) {
+    if (direction !== 'next' && direction !== 'previous') {
+      return { text: 'I don\'t recognize that, type \'next\' or \'previous\' to navigate results, or \'search "query"\' for a new search.' };
+    }
+    const results = await SearchResults.latest(this.meme.get('id'));
+    if (direction === 'next' && results.hasNext()) {
       results.next();
+      winston.info(`getting attachment ${results.get('index')}`);
+      return results.getAttachment();
+    } else if (direction === 'previous' && results.hasPrevious()) {
+      results.preivous();
       return results.getAttachment();
     }
     await this.meme.save({ phase: Phase.QUERY });
@@ -85,19 +104,25 @@ class Memify {
   }
 
   async downloadImage() {
-    const results = await SearchResults.forge({ meme_id: this.meme.id }).fetch({ require: true });
+    const results = await SearchResults.latest(this.meme.get('id'));
     const image = results.get('images')[results.get('index')];
     const url = image['url'];
     const data = await request(url, { encoding: null });
     const filename = `static/templates/cas/${hash(data)}${path.extname(url)}`;
-    fs.writeFileSync(filename, data, { encoding: null });
+    if (!fs.existsSync(filename)) {
+      winston.info(`saving ${url} as ${filename}`);
+      fs.writeFileSync(filename, data, { encoding: null });
+    } else {
+      winston.info(`Using cached copy of ${url}: ${filename}`);
+    }
     await this.meme.save({ template: filename });
   }
 
   async caption() {
+    winston.info(`captioning`);
     const captions = Caption.parse(this.message.text);
     for (const caption of captions) {
-      caption.save({ meme_id: this.meme.id });
+      caption.save({ meme_id: this.meme.get('id') });
     }
     const template = this.meme.get('template');
     const filename = `static/memes/${this.message.team}/${this.message.user}/${hash({ template, captions })}${path.extname(template)}`;
@@ -106,7 +131,8 @@ class Memify {
     return {
       attachments: [
         {
-          title: 'Tada! To recaption use `"your new caption". To start a new search, use \'search "your query terms"\'.',
+          title: `Tada! ${image.url}`,
+          text: 'To recaption type `"your new caption". To start a new search, type \'search "your query terms"\'.',
           image_url: image.url,
         },
       ],
